Possibly ignore SkipShow and remove type checks of metadata sources.
Add new optional argument enable_skip_show to the _generate_feed method which allows you to ignore
SkipShow exceptions from show metadata sources. This new argument is sent to
_populate_show_metadata, which then catches and ignores SkipShow exceptions if enable_skip_show is set to False.

Additionally, remove the assert statements in episode_metadata_sources and show_metadata_sources because
Python is better suited for duck typing, and it's odd to require sources to implement one specific
class in Python (for example, it should be possible to have a source accept both Show and Episode and act
accordingly).