Fix #18 by saving episode sizes in SQLite DB
Save ourselves from 1 roundtrip to the episode host
for each single episode, every single time. Instead,
we ask for the size only the first time, and use the
cached value later.

The only issue is that the db file grows quickly in size.
This could be because long URL strings are used as primary key.
A possible solution would be to create a hash of the URL, and only
save the hash, since it's only used for lookup anyway.