Remove support for running in parallels
Each parallel had to fetch its own list of episodes, metadata and
so on. Therefore, it turned out to be more efficient to do everything
in one single process. (Parallels could only compete when you spawned
20 forks, but then you always had processes that got delayed because
the metadata host would take some time to respond.)

I also believe this is for the best, considering load and stress
for the Radio REST API.