Remove support for --durations option in generate_feed.py and batch_generate_feed.py
A separate script was added in f37e55078ae7644ce761d13cbcaa5d8762e78387 for
calculating episode durations. It is therefore no longer necessary to
support this in the other scripts. The --durations option may trip people
over, since it is not obvious that it will take several hours. It also
introduces different ways of achieving the same thing, which may cause
confusion. It is also bad to mix feed generation and duration calculation,
since the two activities are different in nature.

If you still wish to calculate durations and write feeds in one go, you
may use the && operator in the shell, like this:

python calculate_durations.py && python batch_generate_feed.py ...

This will first calculate durations, and then generate the feed if
calculate_durations.py succeeds.

Issues: It is currently not possible to restrict calculate_durations.py
to only a couple shows; it is "all or nothing".