Store files with write-permission in their own directory by default
The data directory is added. Inside it, files which must be writable
by the program can be stored. By not storing them together with the
executables, you can restrict the program's permissions to write to
the executable files while allowing it to write to the database files.

The .keep file is there so the data directory can be added to git.