Fix #40 so Chimera doens't hide episodes
Before this change, episodes would be hidden if they weren't
present in Chimera. However, the API turned out to be unreliable,
lagging several hours (sometimes half a day!) behind, which meant
episodes would be delayed in the feed for no good reason. With
this change, episodes will appear as soon as they're moved to the
podcast folder, and they'll be supplied with metadata from Chimera
when it surfaces.

This makes it not possible to delay the publication of an
episode, unless we find a way to return future episodes from the
Chimera API.