Stop using feedgen and start using podgen
A bare minimum of changes are done. In the future, I want
to remove the layer that covers podgen, and let the
metadata sources use that framework directly. That should
make the code much more approachable. Fetching duration,
fetching size and saving the outcome of those two should
be separated and put in a separate project, so it can
benefit users of PodGen as well.