Implement utility for populating with old URLs, descripte db init
alternate_show_names.py can be run as a module to use
ALTERNATE_SHOW_NAMES to populate the database. This only needs
to be done once, when populating the empty database. It also
needs for every show to be in the database already; run
py.test webserver/test_api.py to achieve that.