Use custom date for Nerdeprat #057
Digas used the date the podcast was created, but since it
was produced before the publish date, the creation date is
a couple days ago. To avoid users being confused about why an
episode appeared now when it says it has been published several
days ago, I'm changing the date.