Don't use Kapina in pipeline
Kapina does not have a reliable way to search for a show given
information available to podkast.radiorevolt.no. The closest thing
right now is the show title, but it may very likely change in
Kapina, resulting in metadata suddenly disappearing from the podcast
application. Given this, it is not wise to use Kapina metadata in
production, since that would lead to a very fragile dependency.