Improve performance by a lot, catch errors earlier, print warning source
Copying episode and show objects took >99% of the execution time
(discovered using cProfile and making seven calls with the Flask
test client instead of just one, then exiting).
This is fixed by don't caching the objects and create them from the
raw data each and every time (you don't want to process episodes or
shows based on an earlier, processed episode or show).

An attempt is made to handle a request the moment the server starts up,
with any exceptions occurring being passed to the app.run call.
This way, we exercise the init_globals functions and the
machinery around serving feeds when the application starts, meaning
that for example the service will fail immediately if there's a
misconfiguration somewhere (thus, you can look at the service status
rather than the log to determine if we are successful). Just one
call will not exercise all the code, though, so it will still be
necessary to monitor the logs.

Warnings didn't include the caller which triggered the warning.
This is changed so that the filename and linenumber is printed.

In order to not return the same Show instance to two different
requests, it was necessary to do an API change in ShowSource.
The old methods are no longer cached, and yield DeprecationWarning
when invoked.