Implement more explicit and powerful pipeline configuration
You may now embed pipelines as a part of other pipelines.
The settings structure is changed so that each use case
has its own pipeline, and it is up to whoever writes the
configuration to determine what processors should be shared
between different pipelines and how that is done.

The initialized pipelines are now stored in two different
dictionaries, one for shows and one for episodes, and
they are no longer manually assigned a name (if it's a
pipeline in the configuration, it will be available to
the application). There is a check to see if all obligatory
pipelines are configured, though.

I believe this is easier to understand for whoever needs
to touch the configuration, since the inclusion of other
pipelines is done explicitly through the configuration
rather than hard-coded in the source code.