Rename webserver to app, remove unneeded scripts
The name "app" is better at communicating what you'll find inside,
plus it's familiar for those who have gone through tutorials about
Flask.

run.py and run_process_images.py were created to mitigate the
consequences of using relative imports in the application, since
that required the scripts webserver.py (now app.py) and
process_images.py to be called as parts of a package. With the
change to absolute packages, this is no longer necessary.